refactor:  도메인 - 함수 분리 및 상수화 적용
