+++ src/js/domain/auth/__test__/auth.test.js
