+++ src/js/domain/vendingMachine/__test__/moneyBox.test.js
