+++ src/js/utils/dom.js
@@ -12,7 +12,6 @@ export function createMainElement(template) {
 export function createDivElement(template) {
   const divElement = document.createElement('div');
   divElement.id = 'app-inner';
-
   divElement.insertAdjacentHTML('beforeend', template);
 
   return divElement;
